Implemented multithreading to handle logging and hosting the server on a separate thread. This allows the system to gather and log to the sd card while handeling requests from the client.
A mutuexlock has been implemented for access to
the sd card to prevent multiple tasks
(server reading and logging writing)from accessing
the data simultaneously and crashing the system or
corrupting data.

The microcontoller has been changed from the
QTPY ESP32 S2 to the QTPY ESP32 S3 to utilize the
dual core processor.